Add typing to client
