+++ src/@jupiterone/snyk-client.d.ts
@@ -1,1 +1,13 @@
-declare module '@jupiterone/snyk-client';
+declare module '@jupiterone/snyk-client' {
+  export default class SnykModule {
+    constructor(apiKey: string, options?: { retries?: number });
+
+    verifyAccess(orgId: string);
+
+    listAllProjects(orgId: string);
+
+    listIssues(orgId: string, projectId: string, filters?: any);
+
+    listAggregatedIssues(orgId: string, projectId: string, filters?: any);
+  }
+}
